+++ packages/browser/types/index.d.ts
@@ -1,6 +1,3 @@
-import { History, BlockingHistory, LocationComponents, PartialLocation, SessionLocation, AnyLocation, Location, LocationUtilOptions } from "@hickory/root";
-export { History, SessionLocation, PartialLocation, AnyLocation, Location, LocationComponents };
-export declare type Options<Q> = LocationUtilOptions<Q>;
-export declare type BrowserHistory<Q> = History<Q> & BlockingHistory<Q>;
-declare function Browser<Q = string>(options?: Options<Q>): History<Q>;
-export { Browser };
+import { BrowserHistory, Options } from "./types";
+export * from "./types";
+export declare function Browser<Q = string>(options?: Options<Q>): BrowserHistory<Q>;